requirements.dev.txt
fastapi[standard]>=0.115.8
httpx>=0.28.1
pydantic-settings>=2.8.0
pytest>=8.3.5
pytest-asyncio>=0.26.0
pytest-cov>=6.1.1
loguru>=0.7.3
mysql-connector-python>=9.2.0
PyJWT==2.10.1
cryptography>=44.0.3